fixed linter issue
